+++ CHANGELOG.markdown
@@ -48,3 +48,4 @@ Changes in 1.0.0-b3:
 
 * Fix bug with `Request\Response` and IP lookups
 * Fix wrong default value being used for evidence when reporting
+* Added a fourth (optional) parameter to `Library->report()` to allow easy reporting of evidence